+++ sql/create_db.sql
@@ -1,0 +1,5 @@
+DROP DATABASE IF EXISTS todo;
+
+CREATE DATABASE todo;
+
+\c todo