+++ sql/seed_tasks.sql
@@ -1,0 +1,7 @@
+INSERT INTO todo(task)
+
+VALUES
+
+('Pickup groceries'),
+('Wash dishes'),
+('Walk the dog');